added accessibility invalidation
