+++ Source/Protocols/Plugins/Generators/Accessibility/AccessibilityItem.swift
@@ -7,114 +7,9 @@
 
 import UIKit
 
-/// Default accessibility strategy for string value
-public enum AccessibilityStringStrategy {
-
-    /// value will not be change by this strategy
-    case ignored
-
-    /// simple string value, can be nil
-    case just(String?)
-
-    /// a reference value from object. By default it's `accessibilityLabel`
-    /// - parameter keyPath: specified keypath from value would be taken
-    case from(object: NSObject, keyPath: KeyPath<NSObject, String?> = \.accessibilityLabel)
-
-    /// a combination of strategies
-    indirect case joined([AccessibilityStringStrategy])
-
-    var isIgnored: Bool {
-        if case .ignored = self {
-            return true
-        } else {
-            return false
-        }
-    }
-
-    var value: String? {
-        switch self {
-        case .ignored:
-            return nil
-        case .just(let string):
-            return string
-        case .from(let object, let keyPath):
-            return object[keyPath: keyPath]
-        case .joined(let strategies):
-            return strategies.compactMap(\.value).joined()
-        }
-    }
-}
-
-/// Accessibility strategy for `UIAccessibilityTraits` parameter
-public enum AccessibilityTraitsStrategy {
-
-    /// value will not be change by this strategy
-    case ignored
-
-    /// simple accessibility traits
-    case just(UIAccessibilityTraits)
-
-    /// a reference traits from another object
-    case from(object: NSObject)
-
-    /// a reference traits merged from another objects
-    case merge([NSObject])
-
-    var isIgnored: Bool {
-        if case .ignored = self {
-            return true
-        } else {
-            return false
-        }
-    }
-
-    /// accessibility traits of current strategy
-    /// - Returns: nil if `isIgnored`, otherwise `UIAccessibilityTraits`
-    var value: UIAccessibilityTraits? {
-        switch self {
-        case .ignored:
-            return nil
-        case .just(let traits):
-            return traits
-        case .from(let object):
-            return object.accessibilityTraits
-        case .merge(let objects):
-            return objects.map(\.accessibilityTraits).reduce(UIAccessibilityTraits(), { $0.union($1) })
-        }
-    }
-}
-
-/// Accessibility strategies provider protocol for generators
-public protocol AccessibilityStrategyProvider {
-
-    /// strategy for `accessibilityLabel`. Default: `.ignored`
-    var labelStrategy: AccessibilityStringStrategy { get }
-
-    /// strategy for `accessibilityValue`. Default: `.ignored`
-    var valueStrategy: AccessibilityStringStrategy { get }
-
-    /// strategy for `accessibilityTraits`. Default: `.ignored`
-    var traitsStrategy: AccessibilityTraitsStrategy { get }
-
-    /// Idicates that `AccessibilityItem` should become an accessibility element. Equals `true` if all strategies is in state `.ignored`
-    var isAccessibilityIgnored: Bool { get }
-}
-
-public extension AccessibilityStrategyProvider {
-    var labelStrategy: AccessibilityStringStrategy { .ignored }
-    var valueStrategy: AccessibilityStringStrategy { .ignored }
-    var traitsStrategy: AccessibilityTraitsStrategy { .ignored }
-    var isAccessibilityIgnored: Bool {
-        return [labelStrategy.isIgnored, valueStrategy.isIgnored, traitsStrategy.isIgnored].allSatisfy { $0 }
-    }
-}
-
-public protocol AccessibilityActionsProvider {
-    func accessibilityActions() -> [UIAccessibilityCustomAction]
-}
-
 /// Protocol for cells to adopt accesibility
 public protocol AccessibilityItem: UIView, AccessibilityStrategyProvider & AccessibilityActionsProvider {
+    
     typealias AccessibilityModifierType = AccessibilityModifier.Type
 
     /// Type of modifier that will be used to apply strategies to view. Default it's internal implementation
@@ -144,35 +39,3 @@ public extension AccessibilityItem {
     func accessibilityActions() -> [UIAccessibilityCustomAction] { [] }
 }
 
-public enum AccessibilityItemKind {
-    case header(Int), cell(IndexPath), footer(Int)
-}
-
-public struct AccessibilityItemInvalidator {
-    let accessibilityItemKind: AccessibilityItemKind
-    weak var item: AccessibilityItem?
-    weak var accessibilityDelegate: AccessibilityItemDelegate?
-
-    public func invalidateAccessibility() {
-        guard let item else { return }
-        accessibilityDelegate?.didInvalidateAccessibility(for: item, of: accessibilityItemKind)
-    }
-}
-
-public protocol AccessibilityInvalidatable: AccessibilityItem {
-    var accessibilityInvalidator: AccessibilityItemInvalidator? { get set }
-}
-
-extension AccessibilityInvalidatable {
-    func setInvalidator(kind: AccessibilityItemKind, delegate: AccessibilityItemDelegate) {
-        accessibilityInvalidator = AccessibilityItemInvalidator(accessibilityItemKind: kind, item: self, accessibilityDelegate: delegate)
-    }
-
-    func removeInvalidator() {
-        accessibilityInvalidator = nil
-    }
-}
-
-public protocol AccessibilityItemDelegate: AnyObject {
-    func didInvalidateAccessibility(for item: AccessibilityItem, of kind: AccessibilityItemKind)
-}
